% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppr_support.R
\name{add_edge}
\alias{add_edge}
\title{add_edge}
\usage{
add_edge(graph, from, to, l)
}
\arguments{
\item{graph}{a graph constructed with the cppRouting package.}

\item{from}{the node from which to start the edge, has to be in the original nodes.}

\item{to}{the end node of the edge}

\item{l}{the length of the edge.}
}
\value{
the updated graph
}
\description{
Add an edge to a cppRouting network
}
\examples{

# make a small reprex
a <- 1

}
