% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppr_support.R
\docType{data}
\name{cppr}
\alias{cppr}
\title{cppr list of functions.}
\format{
\subsection{\code{cppr}}{

This is an attempt to adapt a module approach to a set of functions in R, inspired by this feature in python.
This is a list object imported with the package and containing support functions for the cppr package.
}
}
\source{
Ivann Schlosser, 2023
}
\usage{
cppr
}
\description{
cppr list of functions.
}
\keyword{datasets}
