% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divbscan.R
\docType{data}
\name{divbscan}
\alias{divbscan}
\title{divbscan list of functions.}
\format{
\subsection{divbscan}{

This is an attempt to adapt a module approach to a set of functions in R, inspired by this feature in python.
This is a list object imported as data with the package and containing support functions for the divbscan algorithm.
}
}
\source{
Ivann Schlosser, 2023
}
\usage{
divbscan
}
\description{
divbscan list of functions.
}
\keyword{datasets}
