% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divbscan.R
\name{entropy_iso}
\alias{entropy_iso}
\title{entropy_iso}
\usage{
entropy_iso(d, iso, by_ = "amenity", cor_num = 2)
}
\arguments{
\item{d}{a data set of spatial features with a categorical variable}

\item{iso}{the accessibility areas for each location of interest, isodists for example.}

\item{by_}{the variable containnig the category}

\item{cor_num}{how many cores to use for greater speed.}
}
\value{
a data frame with columns containing the various variables computed, including the entropy, the size, the ubiquity of the cluster
}
\description{
Computes diversity measure for a set of categorical features contained whithin polygons provided in iso
}
