% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{factor_to_character}
\alias{factor_to_character}
\title{factor_to_character}
\usage{
factor_to_character(dat)
}
\arguments{
\item{dat}{a data frame that contains columns woth factors that we want to turn into character}
}
\value{
the data frame where columns that had factors are turned into character
}
\description{
Turn columns containing factors into character.
Solves the common problem when reading in files with character values that get turned into factor
}
\examples{
dt_fac <- data.frame(letters = factor(c("a","b","c"))
                 ,numbers = c(1,2,3))
class(dt_fac$letters)
dt_char <- factor_to_character(dt_fac)
class(dt_char$letters)
}
