% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppr_support.R
\name{fnearest_nodes}
\alias{fnearest_nodes}
\title{fnearest node}
\usage{
fnearest_nodes(graph, pts, nn = 1, local_crs = NULL, ...)
}
\arguments{
\item{graph}{the graph of interest}

\item{pts}{the point(s) of interest, as a sf object, or a vector with c(lat,lon) structure.}

\item{nn}{how many nearest neighbour points to return for each pts.}

\item{local_crs}{if the coordinates are lonlat, provide a code for a local crs projection. Look up here: https://epsg.io/about}

\item{...}{other parameters to RANN::nn2()}
}
\value{
The corresponding nearest nodes in graph to each pts.
}
\description{
Get the nearest network node to a point with projected coordinates extremely fast thanks to RANN
}
\examples{

#Getting the graph for the french city of lille and a sample of 500 amenities
#from OSM provided with the package.
data("lille_amenities")
data("lille_graph")

loc_crs <- 27561

res <- cppr$fnearest_nodes(lille_graph,lille_amenities,local_crs = loc_crs)

lille_graph$coords[res[[1]],]


}
