% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_lcc}
\alias{get_lcc}
\title{get_lcc}
\usage{
get_lcc(ways, graph_mode = "weak")
}
\arguments{
\item{ways}{a data table containing at least 2 columns 'from' and 'to'.}

\item{graph_mode}{igraph parameter, 'weak' or 'strong' connected component}
}
\value{
a data table that has the same structure as the input, but with removed rows that don't belong to the lcc.
}
\description{
returns the largest connected component of a table of edges and conserves all the attributes. Uses igraph.
examples
}
