% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_lines}
\alias{get_lines}
\title{get_lines}
\usage{
get_lines(from, to = NULL, crs = 4326)
}
\arguments{
\item{from, to}{a matrix with 2 columns containing X and Y coordinates}

\item{crs}{integer number, the code of the reference system}
}
\value{
a sf object with linestrings
}
\description{
This function is intended to turn sets of points into lines
If the to variable is null, then it is assumed that from
is a matrix where each row contains coordinates of a points
the line is then built by connecting each point in the order of the rows
}
\examples{

from <- matrix(data=c(-0.13619,51.52203
                     ,-0.13064,51.52275)
              ,ncol=2
              ,byrow=TRUE)

work_to_lunch <- get_lines(from=from)
plot(work_to_lunch)
}
