% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSMdata.R
\name{getOSMdata}
\alias{getOSMdata}
\title{getOSMdata}
\usage{
getOSMdata(bb, k, val = "all", timeout = 600, memsize = 3073741824, ...)
}
\arguments{
\item{bb}{a bbox}

\item{k}{OSM keys}

\item{val}{OSM values corresponding to the keys}

\item{timeout}{timeout for downloading. increase this value for bigger data sets.}

\item{memsize}{the maximum memory to dedicate to the process. increase this value for bigger data sets.}

\item{...}{other parameters to the osmdata::opq function.}
}
\value{
a list of lists containing the results from the overpass query in different spatial formats.
}
\description{
Facilitated function for working getting data from OSM with osmdata.
}
\examples{

key <- 'amenity'
val <- 'place_of_worship'


}
