% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{index_html}
\alias{index_html}
\title{index_html}
\usage{
index_html()
}
\value{
it is a void function that produces as output the html
file in the same directory in which the rmarkdown file is.
}
\description{
Render a markdown file into a html file called index.html
, for efficient creation of html pages directly publishable to github pages for example.
}
