% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppr_support.R
\docType{data}
\name{lille_amenities}
\alias{lille_amenities}
\title{lille amenities}
\format{
\subsection{lille_amenities}{

A sample of OSM amenities from the city of Lille
}
}
\source{
Ivann Schlosser, 2023
}
\usage{
lille_amenities
}
\description{
lille amenities
}
\keyword{datasets}
