% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppr_support.R
\docType{data}
\name{lille_graph}
\alias{lille_graph}
\title{cppRouting graph for examples.}
\format{
\subsection{\code{lille_graph}}{

A graph constructed from OSM data with the cppRouting package.
}
}
\source{
Ivann Schlosser, 2023
}
\usage{
lille_graph
}
\description{
cppRouting graph for examples.
}
\keyword{datasets}
