% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_distance.R
\docType{data}
\name{london_msoa}
\alias{london_msoa}
\title{london_msoa}
\format{
\subsection{\code{london_msoa}}{

A data set with the msoa of London
}
}
\source{
ONS, 2011
}
\usage{
london_msoa
}
\description{
london_msoa
}
\keyword{datasets}
