% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppr_support.R
\name{make_lines}
\alias{make_lines}
\title{make_lines}
\usage{
make_lines(paths, graph, crs = 4326)
}
\arguments{
\item{paths}{a list of nodes constituting paths returned by get_path_pair, get_multi_paths}

\item{graph}{the graph used.}

\item{crs}{the crs of the data}
}
\value{
an sf linestring data frame
}
\description{
Make an sf linesting data frame out of a set of paths as returned by the cppRouting get_path_pair, get_multi_paths
}
\examples{

# make a small reprex
a <- 1

}
