% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppr_support.R
\name{make_network}
\alias{make_network}
\title{make_network}
\usage{
make_network(edges, nodes = NULL, simple = TRUE, directed = FALSE)
}
\arguments{
\item{edges}{the edges with at least the columns from,to,length.}

\item{nodes}{the nodes with columns id,x,y. Optionnal.}

\item{simple}{simplify the network, default TRUE}

\item{directed}{is the network directed, default FALSE}
}
\value{
The edges data frame of the largest connected component
}
\description{
From a data frame of edges and nodes construct a cppRouting graph.
}
\examples{

# make a small reprex
a <- 1

}
