% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{make_poly}
\alias{make_poly}
\title{make_poly}
\usage{
make_poly(limits, crs_ = 4326)
}
\arguments{
\item{limits}{4 number in the order: x_min,y_min,x_max,y_max}

\item{crs_}{the crs code}
}
\value{
an sf polygon
}
\description{
Function to make a polygon out of 4 numbers representing the limits
}
\examples{

london_bb <- osmdata::getbb("London, UK", limit = 1) |> make_poly()



}
