% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppr_support.R
\name{nearest_nodes}
\alias{nearest_nodes}
\title{nearest node}
\usage{
nearest_nodes(graph, point, crs_ = 4326)
}
\arguments{
\item{graph}{the graph of interest}

\item{point}{the point of interest, as a sf object, or a vector with c(lat,lon) structure.}

\item{crs_}{the crs of the data}
}
\value{
The node with osmid,x,y variables.
}
\description{
Get the nearest network node to a point with lonlat coordinates.
}
\examples{

# make a small reprex
a <- 1

}
