% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppr_support.R
\docType{data}
\name{nodes}
\alias{nodes}
\title{nodes}
\format{
\subsection{\code{nodes}}{

A data frame of nodes to be used with the data set of segments to test the cppr related functions
}
}
\source{
OpenStreetMap, 2024
}
\usage{
nodes
}
\description{
nodes
}
\keyword{datasets}
