% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{overlap}
\alias{overlap}
\title{overlap}
\usage{
overlap(x, y, frac = TRUE)
}
\arguments{
\item{x, y}{a set (vector) of elements}

\item{frac}{logical, TRUE to return the fraction of corresponding elements, FALSE to return the number}
}
\value{
either a values between 0 and 1 for the similarity of the sets
or an integer for the number of common elements
}
\description{
Helper function to for set comparisons. outputs either the fraction (DICE similarity) of
overlap in the sets. Or the number of common elements
}
\details{
Uses the base function intersect and union, which consider only the unique elements of a set.
}
\examples{
x = c(1,2,3,4,5,55,5)
y = c(5,4,3,2,1,1,1,1)
overlap(x,y)
overlap(x,y,frac = FALSE)

}
