% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppr_support.R
\name{plot_path}
\alias{plot_path}
\title{plot_path}
\usage{
plot_path(graph, from_id, to_id)
}
\arguments{
\item{graph}{a graph constructed with the cppRouting package.}

\item{from_id}{the node from which to start the path, has to be in the original nodes.}

\item{to_id}{the end node of the path}
}
\value{
plots the path with tmap
}
\description{
plot a path between two nodes. Useful for debugging and quick visualisations.
}
\examples{

# make a small reprex
a <- 1

}
