% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{samp_dt}
\alias{samp_dt}
\title{samp_dt}
\usage{
samp_dt(dt, weight)
}
\arguments{
\item{dt}{a data.table or data.frame}

\item{weight}{either a number less that 1 for a fraction of the dt,
or a positive integer for a specific number of rows}
}
\value{
a sample of the data set
}
\description{
sample a data table or data frame. Can be used in pipes or on its own.
}
\examples{

data("cars")

cars_half <- samp_dt(cars, .5)
cars_10 <- samp_dt(cars,10)



}
