% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppr_support.R
\docType{data}
\name{segments}
\alias{segments}
\title{segments}
\format{
\subsection{\code{segments}}{

A data frame of segments to be used with the data set of nodes to test the cppr related functions
}
}
\source{
OpenStreetMap, 2024
}
\usage{
segments
}
\description{
segments
}
\keyword{datasets}
