% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{equals}
\alias{equals}
\title{equals}
\usage{
equals(x, y)
}
\arguments{
\item{x, y}{two elements to compare}
}
\value{
a vector of logical values,
with each value corresponding to the comparison between the elements of the entered vectors.
}
\description{
vectorized equal function to use in pipes. Useful for comparing values for summary statistics for example
}
\examples{
x = c(1,2,3,4,5)
y = c(1,3,2,4,5)

x |> equals(y)

}
