% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_lines}
\alias{get_lines}
\title{get_lines}
\usage{
get_lines(from, to = NULL, crs = 4326, by_element = TRUE)
}
\arguments{
\item{from, to}{a matrix with 2 columns containing X and Y coordinates}

\item{crs}{integer number, the code of the reference system}

\item{by_element}{not used in current implementation}
}
\value{
a sf object with linestrings
}
\description{
This function is intended to turn sets of points into lines
If the to variable is null, then it is assumed that from
is a matrix where each row contains coordinates of a points
the line is then built by connecting each point in the order of the rows
}
