% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSMdata.R
\name{osm_group_points}
\alias{osm_group_points}
\title{osm_group_points}
\usage{
osm_group_points(d, k, kofinterest = NULL)
}
\arguments{
\item{d}{a list of lists as returned by getOSMdata}

\item{k}{the keys of interest}

\item{kofinterest}{values of interest for the key k}
}
\value{
a spatial data frame with points.
}
\description{
Function for grouping the outputs of getOSMdata into a single spatial data frame of points. Useful for extracting polygonal or point like features. Polygons are transformed into their centroids.
}
\examples{

key <- 'amenity'
val <- 'place_of_worship'



}
